Updates after review * Move datatypes_struct.py classes to datatypes.py * Make UnsignedN and IntegerN more pythonic * Add comments to decode_raw * Remove commens from encode_raw * Rename all_datatypes.eds to datatypes.eds * Updated test and added comments
